Change sword stats to 0 - 1 floats

+++ Source/Daybreak/DaybreakSword.h
@@ -21,13 +21,13 @@ struct FSwordLevel {
 	FLinearColor Color;
 
     UPROPERTY()
-    int Damage;
+    float Damage;
 
     UPROPERTY()
-    int Speed;
+    float Speed;
 	
 	UPROPERTY()
-    int Range;
+    float Range;
 
     FSwordLevel() {
         Index = 0;
